requirements.txt
slack_bolt
fastapi
uvicorn
httpx
aiohttp
pydantic
psycopg2-binary
pandas
openpyxl
python_dotenv